wiz_level_change cleanup
     The new routine needed to do something with u.ulevelmax
(which controls whether lost levels can be regained via magic).
This sets it to the new level rather than having it go up or
down the same amount as the user specified in the level change,
which yields a plausible reason for using this command to set
experience level explicitly to its current value.

     The new code was using a mixture of spaces and tabs for
indentation; I've totally reformatted it.