airplane/taxi transportation
     The player can teleport objects and monsters on no-teleport
levels, a strange quirk which I think has become entranched as
a feature.  When swallowed or engulfed, teleporting the monster
from inside ends up teleporting the character along with that
monster.  Some players have been exploting this on Plane of Air
to avoid facing elementals and dragons and whatnot by repeatedly
teleporting any vortex that engulfs them until they land somewhere
in the vicinity of the portal leading to Plane of Fire.

     This patch divides the Plane of Air into three zones that
teleportation can't cross.  You'll arrive in the left-hand 30% of
the level, as before, but no longer at a specific spot.  The exit
portal is in the right-hand 30% as before (although it used to
have more range, perhaps 40%).  Teleporting within the left 30%
always arrives in that same area; within the central 40% always
remains within that same area; and teleporting within the right
30% always sticks in that area.  So it's still possible to get
around quite a bit via multiple teleports, but you'll need to walk
at least across the two unmarked boundaries to actually traverse
the whole level.

     A moderately long description for a very short patch....