nttty hilite_pet follow-up
I didn't mean to override for any color in the mask. My
earlier patch caused it to make green, blue, or red all
become black (which still looked better than white on
grey).

This gets it right (famous last words), and only makes
the text character's color black if it is actually white
or bright white.