(from <Someone>) This patch implements Michael's idea to have a "NetHack mode" for keypresses. It only does the Alt-keys now, but it should be easy to extend to pressing space in dialogs; I left that out as we still have to agree on a new dialog look.
The setting that you choose from the menu is stored in the registry.
Reading and writing the registry is done on init_nhwindows and
exit_nhwindows; there also is a menu item to delete the registry
settings (which makes sure they are not stored again this session.)
This provides a framework for future registry settings.

Also,
Alt-tab should show the app title when a menu or text window is up.

I left the app title as it was, although I still favour "Graphical
Version" or "With Graphics"; Anyway, that's only one resource string.