build fix for files.c
Warwick's plname files.c addition broke the
build on both win32 and CE because NAMES_MAX
wasn't defined.

In win32 it was defined in limits.h, but only
when _POSIX_ was defined.

In CE it just didn't exist in any of the
header files.  Since it was also complaining
about strdup(), I #ifdef'd Warwick's code out
under CE.