B13004 spellbook destruction When reading a cursed or too-hard book that's covered in contact poison (presumably in the too-hard case, reading it made the poison), and you die, the book would not be in the bones.  remove in_use mark while assessing damage.
On the other hand, not From a bug report, the message in the "6" case says
the book explodes, but had a 1/3 chance of not disappearing in the normal
case, and 100% chance of remaining if cursed while reading - when the
player survives.  Return a flag to allow the book to be destroyed in this
case.  No work needed for the death case; in_use is set.