B01006 - containers hitting the floor <Someone> noted that if you dropped a box while levitating, nothing would break.  This is true for sacks too, which shouldn't be immune either. Throwing didn't break contained objects either, not mentioned in his report.
Refactored out the code in kick_object that handled damage for objects
contained in normal containers and added calls in hitfloor and throwit.
kick_object still only calls it for boxes, but other callers will call it
for any object letting it decide if damage is required.  BoH objects aren't
affected, since traditionally the inside of the BoH is "somewhere else".