Long ago we received a bug report complaining that hilite_pet on Win32 (tty) wasn't respecting the setting of "use_inverse" (plain "inverse" at the time I think).
In response, we made it respect the setting.  The
"use_inverse" setting is off by default however,
so we've now had about three complaints about
hilite_pet not working.

So I'm changing the default value for win32 tty
to having "use_inverse" set to TRUE.
They can still override it in the config file
that way.