prevent a panic splitbill: no resident shopkeeper Patch was sent in by <Someone> on Sep 8: This bug causes a number of impossible messages (starting with splitbill: no resident shopkeeper??)
Repeat by:

Enter a large shop.
Wish for a large stack of projectiles.
Sell your projectiles and then pick them up again.
Trap shopkeeper against the door with a scroll of earth.
Throw the projectiles at the shopkeeper to anger him.
Move away from the boulder trap and wait for the shopkeeper to leave the
shop.
Throw one of the projectiles at the free space.