gender and size of leaders and nemeses Combo fix based on several reports and additional research. In quest.txt and/or data.base, the Grand Master, Arch Priest, Ixoth, Master Kaen, Nalzok, Scorpius, and the Master Assassin are all referred to as male via reference: his or him.  Ideally, there would be a way to parameterize this in the quest.txt, but I don't see a clear way to do that at this time. For the time being at least, set the M2_MALE flag for these monsters.
The Dark One was referred to via "his" once in the quest.txt.  This case
I was able to modify the offending text rather than forcing a gender.

Orion and Norn are referred to as a giant in data.base but their size did
not correspond.  I left the symbol as S_HUMAN although perhaps it should
be S_GIANT.  Finally, Orion, the Norn, Cyclops and Lord Surtur, as
giant-types, should be able to tear webs.